Modified header function to take additional parameters. Updated landing page. Renamed folders to country names.
