+++ countries/canada/canada.html
@@ -1,0 +1,50 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Canada</title>
+    <link rel="stylesheet" href="/styles/index.css">
+</head>
+<body>
+    <header></header>
+
+    <main>
+        <section class="title__section">
+            <!-- Title -->
+            <!-- Image -->
+        </section>
+
+        <section class="content__section">
+            <div class="content__column">
+                <section class="cityList">
+    
+                </section>
+                
+            </div>
+            
+            <div class="content__column">
+                <section class="landmarkList">
+    
+                </section>
+            </div>
+            
+            <div class="content__column">
+                <section class="citizenList">
+    
+                </section>
+            </div>
+
+            <div class="content__column__images">
+                <!-- Column Images -->
+            </div>
+        </section>
+
+    </main>
+
+    <script type="module" src="/scripts/luke/canada.js"></script>
+    <footer></footer>
+</body>
+</html>
+
