+++ countries/england/england.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
@@ -7,45 +8,50 @@
     <title>Document</title>
     <link rel="stylesheet" href="../../styles/index.css">
 </head>
+
 <body>
-    <header></header>
+    <section class="backdrop">
+        <section class="card">
+            <header></header>
 
-    <main>
-        <h1>Hello World</h1>
-        <section class="title__section">
-            <!-- Title -->
-            <!-- Image -->
+            <main>
+                <section class="title__section">
+                    <!-- Title -->
+                    <!-- Image -->
+                </section>
+
+                <section class="content__section">
+                    <div class="content__column">
+                        <section class="cityList">
+
+                        </section>
+
+                    </div>
+
+                    <div class="content__column">
+                        <section class="landmarkList">
+
+                        </section>
+                    </div>
+
+                    <div class="content__column">
+                        <section class="citizenList">
+
+                        </section>
+                    </div>
+
+                    <div class="content__column__images">
+                        <!-- Column Images -->
+                    </div>
+                </section>
+
+            </main>
+            <footer></footer>
         </section>
+    </section>
 
-        <section class="content__section">
-            <div class="content__column">
-                <section class="cityList">
-    
-                </section>
-                
-            </div>
-            
-            <div class="content__column">
-                <section class="landmarkList">
-    
-                </section>
-            </div>
-            
-            <div class="content__column">
-                <section class="citizenList">
-    
-                </section>
-            </div>
-
-            <div class="content__column__images">
-                <!-- Column Images -->
-            </div>
-        </section>
-
-    </main>
-
-    <footer></footer>
 
     <script type="module" src="../../scripts/england/england.js"></script>
 </body>
+
 </html>
