+++ countries/shared/footer.html
@@ -1,1 +1,36 @@
-<h1>This is in footer.html</h1>
+<div class="upper__footer__div">
+    <div class="footer__column__div">
+        <h2>About Us</h2>
+        <div class="name__container">
+            <p>Mithrandir</p>
+            <p>Telcontar</p>
+            <p>Legolas</p>
+            <p>Gimli</p>
+            <p>Frodo</p>
+            <p>Sam</p>
+            <p>Merry</p>
+            <p>Pippin</p>
+            <p><strike>Boromir</strike></p>
+        </div>
+    </div>
+    <div class="footer__column__div">
+        <h2>Contact Us</h2>
+        <div class="name__container">
+            <a href="https://www.thetolkienforum.com/wiki/Gwaihir#:~:text=Gwaihir%20is%20a%20descendant%20of,the%20bond%20between%20the%20two.">Gwaihir the Windlord</a>
+        </div>
+    </div>
+    <div class="footer__column__div">
+        <h2>Other</h2>
+        <div class="name__container">
+            <p>Hello there!</p>
+            <p>General Kenobi!</p>
+            <p>You Are A Bold One</p>
+        </div>
+    </div>
+</div>
+<br>
+<hr>
+<br>
+<div class="lower__footer__div">
+    <p>Copyright Info &copy;</p>
+</div>
