+++ index.html
@@ -9,7 +9,7 @@
 </head>
 <body>
     <header></header>
-    <a href="countries/luke/canada.html">CLICK</a>
+
     <main>
         <h1>Hello World!</h1>
         <section class="title__section">
@@ -36,3 +36,6 @@
     </main>
 
     <footer></footer>
+
+    <script type="module" src="scripts/index/index.js"></script>
+</body>
