+++ scripts/canada/canadaData.js
@@ -1,0 +1,72 @@
+export const cityObj = [
+    {
+        name: "Vancouver",
+        image: "",
+        population: "675,218",
+        description: "Vancouver, a bustling west coast seaport in British Columbia, is among Canada’s densest, most ethnically diverse cities. A popular filming location, it’s surrounded by mountains, and also has thriving art, theatre and music scenes. Vancouver Art Gallery is known for its works by regional artists, while the Museum of Anthropology houses preeminent First Nations collections.",
+        link: "https://vancouver.ca/",
+    },
+    {
+        name: "Montréal",
+        img: "",
+        population: "1.78 million",
+        description: "Montréal is the largest city in Canada's Québec province. It’s set on an island in the Saint Lawrence River and named after Mt. Royal, the triple-peaked hill at its heart. Its boroughs, many of which were once independent cities, include neighbourhoods ranging from cobblestoned, French colonial Vieux-Montréal – with the Gothic Revival Notre-Dame Basilica at its centre – to bohemian Plateau.",
+        link: "https://montreal.ca/en/",
+    },
+    {
+        name: "Toronto",
+        img: "",
+        population: "2.93 million",
+        description: "Toronto, the capital of the province of Ontario, is a major Canadian city along Lake Ontario’s northwestern shore. It's a dynamic metropolis with a core of soaring skyscrapers, all dwarfed by the iconic, free-standing CN Tower. Toronto also has many green spaces, from the orderly oval of Queen’s Park to 400-acre High Park and its trails, sports facilities and zoo.",
+        link: "https://www.toronto.ca/",
+    },
+]
+
+export const landmarkObj = [
+    {
+        name: "Hotel de Glace",
+        image: "",
+        yearCreated: "January 1, 2001, however it has to be rebuilt every year.",
+        description: "It might be the most beautiful ice hotel of all! The ice is clear and looks like glass, thus creating this amazing sight. Only a few minutes from the center of Quebec City, the hotel meets and greets its guests with a drink served in an ice glass of course. The rooms come with warm sleeping bags, ensuring a cozy rest.",
+        link: "https://www.valcartier.com/fr/hebergement/hotel-de-glace/",
+    },
+    {
+        name: "Banff National Park ",
+        image: "",
+        yearCreated: "November 25, 1885",
+        description: "Rocky Mountain peaks, turquoise glacial lakes, a picture-perfect mountain town and village, abundant wildlife and scenic drives come together in Banff National Park - Canada’s first national park and the flagship of the nation’s park system. Over three million visitors a year make the pilgrimage to the park for a variety of activities including hiking, biking, skiing and camping in some of the world’s most breathtaking mountain scenery. Banff is part of the Canadian Rocky Mountain Parks UNESCO World Heritage Site.",
+        link: "https://www.pc.gc.ca/en/pn-np/ab/banff",
+    },
+    {
+        name: "Grouse Mountain",
+        image: "",
+        yearCreated: "October 1894, named by the first recorded hikers to reach the summit.",
+        description: "Rising 1,250 metres (4,100 feet) above Vancouver and just 15 minutes from the city’s downtown core, is the vast alpine playground of Grouse Mountain.  The iconic Red Skyride gives visitors and locals alike the opportunity to see the jaw-dropping majestic nature of B.C. unfold in front of them during the 8 minute ride from the Valley Station to the summit. As Vancouver’s most popular attraction, Grouse Mountain is the only truly four-season destination in the lower mainland and features a wildlife refuge, vast selection of dining options, thrilling outdoor adventure, summer activities galore, the best in local snow sports…and much more. One thing is for certain: there is no place quite like the Peak of Vancouver.",
+        link: "https://www.grousemountain.com/about",
+    },
+]
+
+export const citizenObj = [
+    {
+        name: "Ryan Gosling",
+        image: "",
+        occupation: "Actor",
+        description: "Though he received no formal acting training, after MMC, Gosling segued into an acting career, appearing on the TV series Young Hercules (1998) and Breaker High (1997), as well as the films The Slaughter Rule (2002), Murder by Numbers (2002), and Remember the Titans (2000). He first attracted serious critical attention with his performance as the Jewish neo-Nazi in the controversial film The Believer (2001), which won the Grand Jury Prize at the 2001 Sundance Film Festival. He was cast in the part by writer-director Henry Bean, who believed that Gosling's strict upbringing gave him the insight to understand the character Danny, whose obsessiveness with the Judaism he was born into turns to hatred. He was nominated for an Independent Spirit Award as Best Male Lead in 2002 for the role and won the Golden Aries award from the Russian Guild of Film Critics.",
+        link: "https://www.imdb.com/name/nm0331516/",
+    },
+    {
+        name: "Wayne Gretzky",
+        image: "",
+        occupation: "Hockey player (retired)",
+        description: "If there is one sport that is synonymous with Canada, it is Ice Hockey. One athlete that made a name for himself in this sport is Wayne Gretzky. He is the most famous and successful hockey player of all time. Wayne holds forty regular-season records, fifteen playoff records, and six All-Star records. Not only was he the leading point-scorer in NHL history, but also the only NHL player to total over 200 points in one season. He accomplished this four times while tallying over 100 points in 16 professional seasons.",
+        link: "https://en.wikipedia.org/wiki/Wayne_Gretzky",
+    },
+    {
+        name: "Justin Trudeau",
+        image: "",
+        occupation: "Prime Minister",
+        description: "Justin Trudeau is Canada’s 23rd Prime Minister. He grew up under the great influence of his father, Pierre Elliot Trudeau, and his mother Margret Trudeau. He is the oldest of three boys and speaks both French and English. This background sparked his passion for public service and shaped his belief that diversity is Canada’s strength. Before diving into politics, Justin served as the Chair of Katimavik, on the board for the Canadian Avalanche Foundation. He was also an advocate for young people and the environment. Since his early youth, he has been encouraging young people to engage with the issues important to them and participate as active citizens.",
+        link: "https://en.wikipedia.org/wiki/Justin_Trudeau",
+    },
+]
+
