+++ styles/footer.css
@@ -1,0 +1,47 @@
+footer {
+    background-color: black;
+    color: white;
+    padding: 2em 1em;
+}
+
+footer h1 {
+    margin: 0;
+    padding: 1em 0;
+}
+
+footer h2 {
+    font-family: "Montserrat";
+    margin:0;
+    padding-bottom: 1em;
+    text-align: center;
+}
+
+.upper__footer__div {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-evenly;
+
+}
+
+.footer__column__div {
+    flex-basis: 30%;
+}
+
+hr {
+    border-color: grey;
+}
+
+.name__container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.name__container p {
+    margin: 8px 0;
+    text-align: left;
+}
+
+.name__container a {
+    color: white;
+}
